// 03_prototype/geocode/src/components/organisms/viewSite/core/types/cardListType.ts
import { LayerType } from "@/components/organisms/viewSite/core/types/layerType"
// CardData 型の定義
export type CardListType = {
  logoImg: string
  text: string
  dangerBadge?: string
  warningBadge?: string
  successBadge?: string
  primaryBadge?: string
  darkBadge?: string
  isShadow: boolean
  shape: string
  isDisplayLayer: boolean
  infoButtonClick: string
  displayButtonClick: string
  orderButtonClick: string
  layer: LayerType
}